rename file in fix img-url calls
