+++ src/layouts/Footer.js
@@ -1,14 +1,14 @@
 import React from "react";
 import { NavLink } from "react-router-dom";
-import "../../assets/styles/variables.scss";
-import Mylogo from "../../assets/logos/chaoss-logo.png";
-import Deilogo from "../../assets/images/DEIlogo.png";
+import "../assets/styles/variables.scss";
+import Mylogo from "../assets/logos/chaoss-logo.png";
+import Deilogo from "../assets/images/DEIlogo.png";
 import ReactReadMoreReadLess from "react-read-more-read-less";
 
-import slack2 from "../../assets/images/slack.png"
-import facebook from "../../assets/images/facebook.png"
-import twitter from "../../assets/images/twitter.png"
-import linkedin from "../../assets/images/linkedin.png"
+import slack2 from "../assets/images/slack.png"
+import facebook from "../assets/images/facebook.png"
+import twitter from "../assets/images/twitter.png"
+import linkedin from "../assets/images/linkedin.png"
 
 
 const Footer = () => {
